requirements.txt
typer[all]==0.12.3
click==8.1.7
rich==13.7.1
pydantic==2.9.2
boto3==1.34.130
pandas==2.2.2
pyarrow==15.0.2
scikit-learn==1.5.1
xgboost==2.1.1
lightgbm==4.3.0
catboost==1.2.5
optuna==3.6.1
mlflow==2.14.1
jinja2==3.1.4
ydata-profiling==4.8.3
psycopg2-binary==2.9.9
SQLAlchemy==2.0.31
redshift-connector==2.1.4
snowflake-connector-python==3.11.0
python-dotenv==1.0.1
ruamel.yaml==0.18.6
openai>=1.40.0
